# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Điện Lạnh Store
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export type ProductSpecs = { [key: string]: string };

export interface Product {
  id: number;
  name: string;
  slug: string;
  brand: string;
  categoryId: number;
  categoryName: string;
  price: number;
  originalPrice?: number;
  discountPercent?: number;
  imageUrl?: string;
  images?: string[];
  description?: string;
  specs?: ProductSpecs;
  isFeatured?: boolean;
  inStock: boolean;
  rating?: number;
  reviewCount?: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl?: string;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddCartItemPayload {
  productId: number;
  quantity: number;
}

export interface UpdateCartQuantityPayload {
  quantity: number;
}

export interface CreateOrderPayload {
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  shippingAddress: string;
  note?: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  shippingAddress: string;
  total: number;
  status: string;
  createdAt: string;
}

export interface AdminOrder {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  shippingAddress: string;
  note?: string;
  total: number;
  status: string;
  createdAt: string;
}

export type UpdateOrderStatusPayloadStatus =
  (typeof UpdateOrderStatusPayloadStatus)[keyof typeof UpdateOrderStatusPayloadStatus];

export const UpdateOrderStatusPayloadStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipping: "shipping",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusPayload {
  status: UpdateOrderStatusPayloadStatus;
}

export type AdminProductPayloadSpecs = { [key: string]: string };

export interface AdminProductPayload {
  name: string;
  slug?: string;
  brand: string;
  categoryId: number;
  price: number;
  originalPrice?: number;
  discountPercent?: number;
  imageUrl?: string;
  description?: string;
  specs?: AdminProductPayloadSpecs;
  isFeatured?: boolean;
  inStock?: boolean;
}

export interface LoginPayload {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Warranty {
  id: number;
  customerName: string;
  phone: string;
  productName: string;
  serialNumber?: string;
  purchaseDate: string;
  warrantyEndDate: string;
  status: string;
  note?: string;
  createdAt: string;
}

export interface CreateWarrantyPayload {
  customerName: string;
  phone: string;
  productName: string;
  serialNumber?: string;
  purchaseDate: string;
  warrantyEndDate: string;
  note?: string;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  brand?: string;
};

export type CheckWarrantyParams = {
  phone: string;
};