# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Điện Lạnh Store
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: cart
    description: Shopping cart operations
  - name: orders
    description: Order operations
  - name: auth
    description: Authentication operations
  - name: admin
    description: Admin operations (protected)
  - name: warranty
    description: Warranty operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      description: List all products, optionally filtered by category
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: brand
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      description: Returns a curated list of featured/best-seller products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      description: Returns all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Current cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemPayload"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartQuantityPayload"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin)
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"

  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusPayload"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"

  /admin/products:
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductPayload"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProductPayload"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/warranties:
    get:
      operationId: adminListWarranties
      tags: [admin]
      summary: List all warranty records (admin)
      responses:
        "200":
          description: All warranties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warranty"
    post:
      operationId: adminCreateWarranty
      tags: [admin]
      summary: Create warranty record (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarrantyPayload"
      responses:
        "201":
          description: Created warranty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"

  /admin/warranties/{id}:
    put:
      operationId: adminUpdateWarranty
      tags: [admin]
      summary: Update warranty record (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarrantyPayload"
      responses:
        "200":
          description: Updated warranty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warranty"
    delete:
      operationId: adminDeleteWarranty
      tags: [admin]
      summary: Delete warranty record (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /warranty/check:
    get:
      operationId: checkWarranty
      tags: [warranty]
      summary: Check warranty by phone number
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Warranty results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warranty"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderPayload"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
      required:
        - id
        - name
        - slug

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        brand:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discountPercent:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        description:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
        isFeatured:
          type: boolean
        inStock:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
      required:
        - id
        - name
        - slug
        - brand
        - categoryId
        - categoryName
        - price
        - inStock

    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - price
        - quantity
        - subtotal

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - total
        - itemCount

    AddCartItemPayload:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity

    UpdateCartQuantityPayload:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity

    CreateOrderPayload:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        note:
          type: string
      required:
        - customerName
        - customerPhone
        - shippingAddress

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - shippingAddress
        - total
        - status
        - createdAt

    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        note:
          type: string
        total:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - shippingAddress
        - total
        - status
        - createdAt

    UpdateOrderStatusPayload:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipping, delivered, cancelled]
      required:
        - status

    AdminProductPayload:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        brand:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        originalPrice:
          type: number
        discountPercent:
          type: integer
        imageUrl:
          type: string
        description:
          type: string
        specs:
          type: object
          additionalProperties:
            type: string
        isFeatured:
          type: boolean
        inStock:
          type: boolean
      required:
        - name
        - brand
        - categoryId
        - price

    LoginPayload:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - role

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Warranty:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        productName:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
        warrantyEndDate:
          type: string
        status:
          type: string
        note:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - phone
        - productName
        - purchaseDate
        - warrantyEndDate
        - status
        - createdAt

    CreateWarrantyPayload:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        productName:
          type: string
        serialNumber:
          type: string
        purchaseDate:
          type: string
        warrantyEndDate:
          type: string
        note:
          type: string
      required:
        - customerName
        - phone
        - productName
        - purchaseDate
        - warrantyEndDate