# lib/api-zod/src/generated/types/adminProductPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Điện Lạnh Store
 * OpenAPI spec version: 0.1.0
 */
import type { AdminProductPayloadSpecs } from "./adminProductPayloadSpecs";

export interface AdminProductPayload {
  name: string;
  slug?: string;
  brand: string;
  categoryId: number;
  price: number;
  originalPrice?: number;
  discountPercent?: number;
  imageUrl?: string;
  description?: string;
  specs?: AdminProductPayloadSpecs;
  isFeatured?: boolean;
  inStock?: boolean;
}