# lib/api-zod/src/generated/types/loginPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Điện Lạnh Store
 * OpenAPI spec version: 0.1.0
 */

export interface LoginPayload {
  username: string;
  password: string;
}