# lib/api-zod/src/generated/types/updateOrderStatusPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Điện Lạnh Store
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateOrderStatusPayloadStatus } from "./updateOrderStatusPayloadStatus";

export interface UpdateOrderStatusPayload {
  status: UpdateOrderStatusPayloadStatus;
}