# lib/api-zod/src/generated/types/warranty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Điện Lạnh Store
 * OpenAPI spec version: 0.1.0
 */

export interface Warranty {
  id: number;
  customerName: string;
  phone: string;
  productName: string;
  serialNumber?: string;
  purchaseDate: string;
  warrantyEndDate: string;
  status: string;
  note?: string;
  createdAt: string;
}